requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
WTForms==3.1.0
Werkzeug==2.3.7
Jinja2==3.1.2
PyCryptodome==3.19.0
reportlab==4.0.7 